test: move mocks so they are co located with queries
